+++ staging/usr/local/opnsense/scripts/ml/models/stats.py
@@ -5,7 +5,7 @@ request_logger = logging.getLogger('django.request')
 
 
 class Stats(models.Model):
-    timestamp = models.IntegerField(default=0)
+    timestamp = models.FloatField(default=0.0)
     files = models.IntegerField(default=0)
 
     def __init__(self, *args, **kwargs):
